move Handshake from NTS.ACL into NTS.Application
